# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "FridgeProduct": {
        "description": "Продукт в холодильнике (прогноз).",
        "properties": {
          "consumption_rate": {
            "title": "Consumption Rate",
            "type": "number"
          },
          "days_until_empty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Until Empty"
          },
          "estimated_quantity": {
            "title": "Estimated Quantity",
            "type": "number"
          },
          "expected_expiry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Expiry"
          },
          "last_purchased": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Purchased"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "product_name",
          "estimated_quantity",
          "unit",
          "last_purchased",
          "expected_expiry",
          "consumption_rate",
          "days_until_empty"
        ],
        "title": "FridgeProduct",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NutritionSummary": {
        "description": "Агрегированные данные о питании за период.",
        "properties": {
          "avg_daily_calories": {
            "title": "Avg Daily Calories",
            "type": "number"
          },
          "avg_daily_carbs": {
            "title": "Avg Daily Carbs",
            "type": "number"
          },
          "avg_daily_fats": {
            "title": "Avg Daily Fats",
            "type": "number"
          },
          "avg_daily_proteins": {
            "title": "Avg Daily Proteins",
            "type": "number"
          },
          "by_category": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "By Category",
            "type": "array"
          },
          "period_from": {
            "title": "Period From",
            "type": "string"
          },
          "period_to": {
            "title": "Period To",
            "type": "string"
          },
          "total_calories": {
            "title": "Total Calories",
            "type": "number"
          },
          "total_carbs": {
            "title": "Total Carbs",
            "type": "number"
          },
          "total_fats": {
            "title": "Total Fats",
            "type": "number"
          },
          "total_proteins": {
            "title": "Total Proteins",
            "type": "number"
          }
        },
        "required": [
          "period_from",
          "period_to",
          "total_calories",
          "total_proteins",
          "total_fats",
          "total_carbs",
          "avg_daily_calories",
          "avg_daily_proteins",
          "avg_daily_fats",
          "avg_daily_carbs"
        ],
        "title": "NutritionSummary",
        "type": "object"
      },
...
      "ProductMatchRequest": {
        "description": "Запрос на распознавание продукта.",
        "properties": {
          "quantity": {
            "default": 1,
            "title": "Quantity",
            "type": "number"
          },
          "raw_name": {
            "title": "Raw Name",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "raw_name"
        ],
        "title": "ProductMatchRequest",
        "type": "object"
      },
...
      "ProductSearchResult": {
        "description": "Результат поиска продуктов.",
        "properties": {
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ProductSchema"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "results"
        ],
        "title": "ProductSearchResult",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Food Spend Tracker API",
    "version": "1.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analytics/nutrition": {
      "get": {
        "description": "Анализ КБЖУ и нутриентов за период.",
        "operationId": "nutrition_analysis_api_analytics_nutrition_get",
        "parameters": [
          {
            "in": "query",
            "name": "from_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "in": "query",
            "name": "to_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Nutrition Analysis",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
    "/api/analytics/spending": {
      "get": {
        "description": "Анализ трат за период.",
        "operationId": "spending_analysis_api_analytics_spending_get",
        "parameters": [
          {
            "in": "query",
            "name": "from_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "in": "query",
            "name": "to_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Spending Analysis",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
...
    "/api/fridge": {
      "get": {
        "description": "Прогноз остатков в холодильнике.",
        "operationId": "fridge_status_api_fridge_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FridgeProduct"
                  },
                  "title": "Response Fridge Status Api Fridge Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fridge Status",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
    "/api/products": {
      "post": {
        "description": "Создание нового продукта (вручную).",
        "operationId": "create_product_api_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Product",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
    "/api/products/match": {
      "post": {
        "description": "Распознавание продукта по сырому названию.",
        "operationId": "match_product_endpoint_api_products_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductMatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductMatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Match Product Endpoint",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
    "/api/products/search": {
      "get": {
        "description": "Поиск продуктов по названию (точный + нечеткий).",
        "operationId": "search_products_api_products_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSearchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Products",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
    "/api/products/{product_id}": {
      "get": {
        "description": "Получение продукта по ID.",
        "operationId": "get_product_api_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Product",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
    "/api/products/{product_id}/substitutes": {
      "get": {
        "description": "Получение аналогов продукта.",
        "operationId": "get_substitutes_api_products__product_id__substitutes_get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductSubstituteSchema"
                  },
                  "title": "Response Get Substitutes Api Products  Product Id  Substitutes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Substitutes",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
...
    "/api/recipes": {
      "get": {
        "description": "Список рецептов.",
        "operationId": "list_recipes_api_recipes_get",
        "parameters": [
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecipeSchema"
                  },
                  "title": "Response List Recipes Api Recipes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recipes",
        "tags": [
          "Knowledge Base"
        ]
      },
      "post": {
        "description": "Создание нового рецепта.",
        "operationId": "create_recipe_api_recipes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Recipe",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
    "/api/recipes/suggest": {
      "get": {
        "description": "\"Что приготовить?\" — подбор рецептов по продуктам пользователя.",
        "operationId": "suggest_recipes_endpoint_api_recipes_suggest_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecipeSuggestion"
                  },
                  "title": "Response Suggest Recipes Endpoint Api Recipes Suggest Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Suggest Recipes Endpoint",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
    "/api/recipes/{recipe_id}": {
      "get": {
        "description": "Получение рецепта по ID.",
        "operationId": "get_recipe_api_recipes__recipe_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "title": "Recipe Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recipe",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
...
  }
}